Update toDispWaves.m
So, can detect all circles and triangles. Horns will not test correctly, due to a small* issue with the data. The 1.5million samples are actually 150k samples, but they hold for 10 seconds. This is a bit annoying to account for, so I'm not going to. The general principle holds. 

Calculating the width from this is also a bit inconvenient. So, I'm gonna leave that out of the matlab code.