Create waveAnalysis_ausf_a.c
Note: Not integrated into .h file. 
Also, note, not tested or shizz. 
Need to establish actual threshold values for this to work. Otherwise it is shit. Will need to have a 'set up' which finds a zero point for the system, then sets the thresholds based off that. Will require testing.


There are no guarantees, written or implied, for the working of this file.